Added a basic preferences page, plus name fields to users and more friendly messages to the home page using the name fields. Also added a cucumber test to check the feature. Finally, removed the old pdfkit test pages
