+++ app/assets/javascripts/users/index.js
@@ -1,0 +1,2 @@
+//
+//= require users/users
