+++ app/assets/javascripts/users/users.js
@@ -1,0 +1,2 @@
+app.users = {
+}
