+++ app/assets/stylesheets/active_admin.css.scss
@@ -1,5 +1,3 @@
-saraba {}
-
 // Active Admin CSS Styles
 @import "active_admin/mixins";
 @import "active_admin/base";
