+++ app/assets/stylesheets/application.css.scss
@@ -24,6 +24,8 @@
 @import "reports/new_edit";
 @import "reports/show";
 
+@import "users/edit";
+
 @import "layouts/header";
 
 @import "static_pages/welcome";
