+++ app/assets/stylesheets/reports/_new_edit.css.scss
@@ -29,4 +29,3 @@
     top: 10px;
   }
 }
-
