+++ app/views/home/index.html.erb
@@ -3,7 +3,7 @@
 
 <div class="container">
 
-  <h1>Welcome back, <%= @user.email %>!</h1>
+  <h1>Welcome back, <%= @user.first_name %>!</h1>
   <h3>Select a report to start communicating!</h3>
 
   <div class="reports-link-container">
@@ -32,4 +32,3 @@
   </div>
 
 </div>
-
