--- app/views/home/pdfkit.html.erb
@@ -1,6 +1,0 @@
-<form action="/pdf">
-  <label>Input URL to transform to PDF here</label>
-  <textarea rows="2" cols="80" id="url" name="url">http://www.google.com</textarea><br/>
-  <br/>
-  <input type="submit" value="Transform to PDF">
-</form>
