+++ app/views/layouts/_header.html.erb
@@ -5,13 +5,16 @@
   <div class="user-menu dropdown">
     <i class=" icon-chevron-down icon-white"></i>
     <a href="#" class="dropdown-toggle" data-toggle="dropdown">
-      <%= current_user.email %>
+      <%= current_user.full_name %>
     </a>
 
     <ul class="dropdown-menu" role="menu" aria-labelledby="drop3">
       <li><a tabindex="-1" href="<%= reports_url %>">My Reports</a></li>
       <li class="divider"></li>
-      <li><a tabindex="-1" href="#">Preferences</a></li>
+      <li>
+        <!--<a tabindex="-1" href="<%= url_for current_user %>">Preferences</a>-->
+        <%= link_to 'Preferences', { :action => :edit, :id => current_user, :controller => "users" }, { :tabindex => "-1" } %>
+      </li>
     </ul>
   </div>
