+++ app/views/users/edit.html.erb
@@ -1,0 +1,29 @@
+<% provide(:title, 'Preferences') %>
+<% provide(:page_id, "users-edit-page") %>
+
+<div class="users-edit-box">
+
+  <%= form_for @user do |f| %>
+
+    <legend>Update your personal information and press Save to continue</legend>
+
+    <%= f.label :last_name %>
+    <%= f.text_field :last_name %>
+
+    <%= f.label :first_name %>
+    <%= f.text_field :first_name %>
+
+    <%= f.label :email %>
+    <%= f.email_field :email %>
+
+    <%= f.label :password %>
+    <%= f.password_field :password %>
+
+    <%= f.label :password_confirmation %>
+    <%= f.password_field :password_confirmation %>
+
+    <%= f.submit "Save", class: "button green-button" %>
+
+  <% end %>
+
+</div>
