+++ config/database.yml
@@ -13,20 +13,17 @@ development:
 # Warning: The database defined as "test" will be erased and
 # re-generated from your development database when you run "rake".
 # Do not set this db to the same as development or production.
-test: &test
-  adapter: sqlite3
-  database: db/test.sqlite3
-  pool: 5
-  timeout: 5000
-#  adapter: postgresql
-#  username: postgres
-#  password: password
-#  database: pretty_reports_test
+test: &test  
+  adapter: postgresql
+  username: postgres
+  password: password
+  database: pretty_reports_test
 
 production:
   adapter: postgresql
   username: prprod
   password: prprod
   database: pretty_reports_production
+
 cucumber:
   <<: *test
