+++ config/routes.rb
@@ -6,8 +6,6 @@ PrettyReports::Application.routes.draw do
   devise_scope :user do get '/users/sign_out' => 'devise/sessions#destroy' end
 
   root to: "home#index"
-  match "/pdfkit", :to => "home#pdfkit"
-  match "/pdf", :to => "home#pdf"
 
   get '/static-pages/about-us' => 'static_pages#about_us'
   get '/static-pages/welcome' => 'static_pages#welcome', as: 'welcome_page'
@@ -18,5 +16,7 @@ PrettyReports::Application.routes.draw do
       get 'download_available', path: 'download-available'
     end
   end
+
+  resources :users
 
 end
