--- db/migrate/20120815050438_add_admin_to_users.rb
@@ -1,5 +1,0 @@
-class AddAdminToUsers < ActiveRecord::Migration
-  def change
-    add_column :users, :admin, :boolean
-  end
-end
