+++ db/migrate/20120815050438_add_fields_to_users.rb
@@ -1,0 +1,7 @@
+class AddFieldsToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :admin, :boolean
+    add_column :users, :first_name, :string
+    add_column :users, :last_name, :string
+  end
+end
