+++ db/schema.rb
@@ -76,6 +76,8 @@ ActiveRecord::Schema.define(:version => 20120924145450) do
     t.datetime "created_at",                             :null => false
     t.datetime "updated_at",                             :null => false
     t.boolean  "admin"
+    t.string   "first_name"
+    t.string   "last_name"
   end
 
   add_index "users", ["email"], :name => "index_users_on_email", :unique => true
