+++ spec/models/user_spec.rb
@@ -16,6 +16,8 @@
 #  created_at             :datetime        not null
 #  updated_at             :datetime        not null
 #  admin                  :boolean
+#  first_name             :string(255)
+#  last_name              :string(255)
 #
 
 require 'spec_helper'
